requirements.txt
Flask~=1.1.2
blinker~=1.4
Flask-ReCaptcha~=0.4.2
SQLAlchemy~=1.3.23
tzlocal~=2.1
sentry_sdk~=0.20.3
Flask-SQLAlchemy~=2.4.4
Flask-Cors~=3.0.10
Flask-Classful~=0.14.2
marshmallow-sqlalchemy~=0.24.2
marshmallow-utils~=0.3.7
psycopg2-binary~=2.8.6
mysqlclient~=2.0.3